removed getAllParks from search.js and moved to homeRoutes.js
